+++ customer.go
@@ -21,29 +21,29 @@ type Customer struct {
 // CreateCustomer creates a customer
 func CreateCustomer(data *Customer) map[string]interface{} {
 	endpoint := APIBaseURL + "customers/create"
-	return Response(data, endpoint)
+	return makeHTTPRequest(data, endpoint)
 }
 
 // AllCustomers retrieves a list of customers
 func AllCustomers(data *Customer) map[string]interface{} {
 	endpoint := APIBaseURL + "customers"
-	return Response(data, endpoint)
+	return makeHTTPRequest(data, endpoint)
 }
 
 // RetrieveCustomer retrieves a single customer
 func RetrieveCustomer(data *Customer) map[string]interface{} {
 	endpoint := APIBaseURL + "customers/" + strconv.Itoa(data.ID)
-	return Response(data, endpoint)
+	return makeHTTPRequest(data, endpoint)
 }
 
 // UpdateCustomer updates a customer
 func UpdateCustomer(data *Customer) map[string]interface{} {
 	endpoint := APIBaseURL + "customers/" + strconv.Itoa(data.ID) + "/update"
-	return Response(data, endpoint)
+	return makeHTTPRequest(data, endpoint)
 }
 
 // DeleteCustomer deletes a customer
 func DeleteCustomer(data *Customer) map[string]interface{} {
 	endpoint := APIBaseURL + "customers/" + strconv.Itoa(data.ID) + "/delete"
-	return Response(data, endpoint)
+	return makeHTTPRequest(data, endpoint)
 }